# refactor: implement append-only tree