--- a/Cargo.toml
+++ b/Cargo.toml
@@ -21,8 +21,6 @@ default = ["std"]
 std = []
 # SMT implemented in C
 smtc = []
-# A storage optimized SMT implemented in trie (https://ouvrard-pierre-alain.medium.com/sparse-merkle-tree-86e6e2fc26da)
-trie = []
 
 [dependencies]
 cfg-if = "1"