--- a/src/h256.rs
+++ b/src/h256.rs
@@ -28,7 +28,7 @@ impl H256 {
     pub fn set_bit(&mut self, i: u8) {
         let byte_pos = i / BYTE_SIZE;
         let bit_pos = i % BYTE_SIZE;
-        self.0[byte_pos as usize] |= 1 << bit_pos as u8;
+        self.0[byte_pos as usize] |= 1 << bit_pos;
     }
 
     #[inline]