--- a/src/lib.rs
+++ b/src/lib.rs
@@ -73,18 +73,13 @@ pub mod merkle_proof;
 mod tests;
 pub mod traits;
 mod tree;
-#[cfg(feature = "trie")]
-mod trie_tree;
 
 #[cfg(feature = "smtc")]
 pub use ckb_smt::{SMTBuilder, SMT};
 pub use h256::H256;
 pub use merkle_proof::{CompiledMerkleProof, MerkleProof};
-#[cfg(not(feature = "trie"))]
+pub use tree::BranchNode;
 pub use tree::SparseMerkleTree;
-pub use tree::{BranchKey, BranchNode};
-#[cfg(feature = "trie")]
-pub use trie_tree::SparseMerkleTree;
 
 /// Expected path size: log2(256) * 2, used for hint vector capacity
 pub const EXPECTED_PATH_SIZE: usize = 16;