--- a/src/merge.rs
+++ b/src/merge.rs
@@ -11,12 +11,7 @@ pub enum MergeValue {
         base_node: H256,
         zero_bits: H256,
         zero_count: u8,
-    },
-    #[cfg(feature = "trie")]
-    ShortCut {
-        key: H256,
         value: H256,
-        height: u8,
     },
 }
 
@@ -33,23 +28,20 @@ impl MergeValue {
         match self {
             MergeValue::Value(v) => v.is_zero(),
             MergeValue::MergeWithZero { .. } => false,
-            #[cfg(feature = "trie")]
-            MergeValue::ShortCut { .. } => false,
         }
     }
 
-    #[cfg(feature = "trie")]
-    pub fn shortcut_or_value(key: H256, value: H256, height: u8) -> Self {
-        if height == 0 || value.is_zero() {
-            MergeValue::Value(value)
-        } else {
-            MergeValue::ShortCut { key, value, height }
+    pub fn dec_zero(&mut self, height: u8) {
+        if let MergeValue::MergeWithZero {
+            base_node: _,
+            zero_bits,
+            zero_count,
+            value: _,
+        } = self
+        {
+            zero_bits.clear_bit(height);
+            *zero_count -= 1;
         }
-    }
-
-    #[cfg(feature = "trie")]
-    pub fn is_shortcut(&self) -> bool {
-        matches!(self, MergeValue::ShortCut { .. })
     }
 
     pub fn hash<H: Hasher + Default>(&self) -> H256 {
@@ -59,6 +51,7 @@ impl MergeValue {
                 base_node,
                 zero_bits,
                 zero_count,
+                value: _,
             } => {
                 let mut hasher = H::default();
                 hasher.write_byte(MERGE_ZEROS);
@@ -67,34 +60,6 @@ impl MergeValue {
                 hasher.write_byte(*zero_count);
                 hasher.finish()
             }
-            #[cfg(feature = "trie")]
-            MergeValue::ShortCut { key, value, height } => {
-                into_merge_value::<H>(*key, *value, *height).hash::<H>()
-            }
-        }
-    }
-}
-
-/// Helper function for Shortcut node
-/// Transform it into a MergeValue or MergeWithZero node
-#[cfg(feature = "trie")]
-pub fn into_merge_value<H: Hasher + Default>(key: H256, value: H256, height: u8) -> MergeValue {
-    // try keep hash same with MergeWithZero
-    if value.is_zero() || height == 0 {
-        MergeValue::from_h256(value)
-    } else {
-        let base_key = key.parent_path(0);
-        let base_node = hash_base_node::<H>(0, &base_key, &value);
-        let mut zero_bits = key;
-        for i in height..=core::u8::MAX {
-            if key.get_bit(i) {
-                zero_bits.clear_bit(i);
-            }
-        }
-        MergeValue::MergeWithZero {
-            base_node,
-            zero_bits,
-            zero_count: height,
         }
     }
 }
@@ -139,6 +104,33 @@ pub fn merge<H: Hasher + Default>(
     MergeValue::Value(hasher.finish())
 }
 
+/// Merge two hash with node information
+/// this function optimized for ZERO_HASH
+/// if lhs and rhs both are ZERO_HASH return ZERO_HASH, otherwise hash all info.
+pub fn _merge<H: Hasher + Default>(
+    height: u8,
+    node_key: &H256,
+    lhs: &MergeValue,
+    rhs: &MergeValue,
+) -> MergeValue {
+    if lhs.is_zero() && rhs.is_zero() {
+        return MergeValue::zero();
+    }
+    if lhs.is_zero() {
+        return merge_with_zero::<H>(height, node_key, rhs, true);
+    }
+    if rhs.is_zero() {
+        return merge_with_zero::<H>(height, node_key, lhs, false);
+    }
+    let mut hasher = H::default();
+    hasher.write_byte(MERGE_NORMAL);
+    hasher.write_byte(height);
+    hasher.write_h256(node_key);
+    hasher.write_h256(&lhs.hash::<H>());
+    hasher.write_h256(&rhs.hash::<H>());
+    MergeValue::Value(hasher.finish())
+}
+
 pub fn merge_with_zero<H: Hasher + Default>(
     height: u8,
     node_key: &H256,
@@ -156,12 +148,14 @@ pub fn merge_with_zero<H: Hasher + Default>(
                 base_node,
                 zero_bits,
                 zero_count: 1,
+                value: *v,
             }
         }
         MergeValue::MergeWithZero {
             base_node,
             zero_bits,
             zero_count,
+            value,
         } => {
             let mut zero_bits = *zero_bits;
             if set_bit {
@@ -171,24 +165,57 @@ pub fn merge_with_zero<H: Hasher + Default>(
                 base_node: *base_node,
                 zero_bits,
                 zero_count: zero_count.wrapping_add(1),
+                value: *value,
             }
         }
-        #[cfg(feature = "trie")]
-        MergeValue::ShortCut { key, value, .. } => {
-            if height == core::u8::MAX {
-                let base_key = key.parent_path(0);
-                let base_node = hash_base_node::<H>(0, &base_key, value);
-                MergeValue::MergeWithZero {
-                    base_node,
-                    zero_bits: *key,
-                    zero_count: 0,
+    }
+}
+
+/// Merge node with zeros
+pub fn merge_with_zeros<H: Hasher + Default>(
+    key: H256,
+    value: MergeValue,
+    target_height: u8,
+    zeros: u8,
+) -> MergeValue {
+    if target_height == 0 || zeros == 0 || value.is_zero() {
+        return value;
+    }
+    let base_height = target_height - zeros;
+    match value {
+        MergeValue::Value(value) => {
+            let base_key = key.parent_path(base_height);
+            let base_node = hash_base_node::<H>(base_height, &base_key, &value);
+            let mut zero_bits = H256::zero();
+            for i in base_height..target_height {
+                if key.get_bit(i) {
+                    zero_bits.set_bit(i);
                 }
-            } else {
-                MergeValue::ShortCut {
-                    key: *key,
-                    value: *value,
-                    height: height + 1,
+            }
+            MergeValue::MergeWithZero {
+                base_node,
+                zero_bits,
+                zero_count: zeros,
+                value,
+            }
+        }
+        MergeValue::MergeWithZero {
+            base_node,
+            zero_bits,
+            zero_count,
+            value,
+        } => {
+            let mut zero_bits = zero_bits;
+            for i in base_height..target_height {
+                if key.get_bit(i) {
+                    zero_bits.set_bit(i);
                 }
+            }
+            MergeValue::MergeWithZero {
+                base_node,
+                zero_bits,
+                zero_count: zero_count.wrapping_add(zeros),
+                value,
             }
         }
     }