--- a/src/tests/mod.rs
+++ b/src/tests/mod.rs
@@ -1,4 +1,6 @@
 // FIXME: fix fixtures tests later
 mod fixtures;
+mod history;
 mod smt;
 mod tree;
+mod utils;