--- a/src/traits.rs
+++ b/src/traits.rs
@@ -1,8 +1,4 @@
-use crate::{
-    error::Error,
-    tree::{BranchKey, BranchNode},
-    H256,
-};
+use crate::{error::Error, tree::BranchNode, H256};
 
 /// Trait for customize hash function
 pub trait Hasher {
@@ -28,13 +24,13 @@ impl Value for H256 {
 
 /// Traits for customize backend storage
 pub trait StoreReadOps<V> {
-    fn get_branch(&self, branch_key: &BranchKey) -> Result<Option<BranchNode>, Error>;
+    fn get_branch(&self, branch_key: &H256) -> Result<Option<BranchNode>, Error>;
     fn get_leaf(&self, leaf_key: &H256) -> Result<Option<V>, Error>;
 }
 
 pub trait StoreWriteOps<V> {
-    fn insert_branch(&mut self, node_key: BranchKey, branch: BranchNode) -> Result<(), Error>;
+    fn insert_branch(&mut self, node_key: H256, branch: BranchNode) -> Result<(), Error>;
     fn insert_leaf(&mut self, leaf_key: H256, leaf: V) -> Result<(), Error>;
-    fn remove_branch(&mut self, node_key: &BranchKey) -> Result<(), Error>;
+    fn remove_branch(&mut self, node_key: &H256) -> Result<(), Error>;
     fn remove_leaf(&mut self, leaf_key: &H256) -> Result<(), Error>;
 }